fix: address multiple code review issues across hot paths and utilities
Bug fixes:
- fix(pareto): hypervolume formula now uses ref_acc correctly (was ignored)
- fix(debug_telemetry): handle non-contiguous grads with flatten() not view()
- fix(debug_telemetry): detect NaN/Inf ratios explicitly (NaN comparisons are False)
- fix(debug_telemetry): cap diagnostic payload to 100 exemplars to prevent explosion
- fix(config): broken docstring import path (esper.simic.config → esper.simic.training)
- fix(config): add task field to to_train_kwargs() when set
- fix(config): reject string slots with helpful error message

Performance improvements:
- perf(advantages): remove .float() to preserve dtype under AMP (float16/bfloat16)
- perf(advantages): return (advantages, masks) tuple to eliminate redundant computation
- perf(ppo): remove .float() from KL, policy loss, and entropy loss calculations

API improvements:
- refactor(tolaria): centralize device validation with parse_device() and validate_device()
- refactor(vectorized): use tolaria's validate_device(), removing ~32 lines of duplication

Documentation fixes:
- docs(advantages): add tempo_head to decision tree docstring
- docs(pareto): fix dominance definition (≥ all + > one, not strictly > all)
- docs(causal_masks): update example to show bool multiplication pattern

Tests added:
- test(pareto): 6 new tests for non-zero ref_acc edge cases
- test(advantages): 4 new tests for dtype preservation (float16/bfloat16)
- test(debug_telemetry): tests for non-contiguous grads, NaN detection, payload capping
- test(config): 7 prevention tests for docstring, task field, slots validation
- test(environment): 14 new tests for device validation edge cases

+++ src/esper/leyline/causal_masks.py
@@ -37,10 +37,10 @@ def compute_causal_masks(op_actions: torch.Tensor) -> dict[str, torch.Tensor]:
 
     Example:
         >>> masks = compute_causal_masks(op_actions)
-        >>> # For advantage masking:
-        >>> masked_adv = advantages * masks["blueprint"].float()
+        >>> # For advantage masking (bool multiplication preserves dtype):
+        >>> masked_adv = advantages * masks["blueprint"]
         >>> # For masked mean:
-        >>> head_loss = (loss * masks["slot"].float()).sum() / masks["slot"].sum()
+        >>> head_loss = (loss * masks["slot"]).sum() / masks["slot"].sum()
     """
     is_wait = op_actions == LifecycleOp.WAIT
     is_germinate = op_actions == LifecycleOp.GERMINATE
